+++ src/ifs/wl_seat/event_handling.rs
@@ -196,7 +196,20 @@ impl WlSeatGlobal {
                 time_usec,
                 key,
                 state,
-            } => self.key_event(time_usec, key, state),
+            } => {
+                let desired_kb_map_id = match dev.keymap_id.get() {
+                    Some(id) => id,
+                    None => self.seat_kb_map_id.get(),
+                };
+                if desired_kb_map_id != self.effective_kb_map_id.get() {
+                    let map = match dev.keymap.get() {
+                        Some(map) => map,
+                        None => self.seat_kb_map.get(),
+                    };
+                    self.set_effective_keymap(&map);
+                }
+                self.key_event(time_usec, key, state)
+            }
             InputEvent::ConnectorPosition {
                 time_usec,
                 connector,
@@ -742,10 +755,10 @@ impl WlSeatGlobal {
 // Focus callbacks
 impl WlSeatGlobal {
     pub fn focus_surface(&self, surface: &WlSurface) {
-        let pressed_keys: Vec<_> = self.pressed_keys.borrow().iter().copied().collect();
+        let pressed_keys = &*self.pressed_keys.borrow();
         let serial = surface.client.next_serial();
         self.surface_kb_event(Version::ALL, surface, |k| {
-            k.send_enter(serial, surface.id, &pressed_keys)
+            k.send_enter(serial, surface.id, pressed_keys)
         });
         let ModifierState {
             mods_depressed,